+++ lathe_gui/gui/bsp.h
@@ -3,11 +3,12 @@
  * These are functions provided by HAL that are called by the
  * portable business logic of the program.
  */
-
+#include <inttypes.h>
+#include <stdbool.h>
 
 /**
  * Get the current time in milliseconds since an arbitrary epoch.
  * (On uC epoch is power on, on PC epoch is Unix epoch)
  */
-uint64_t getCurrentTimeMillis();
+uint32_t getCurrentTimeMillis();
